+++ lib/pages/detection/annotation.dart
@@ -1,8 +1,6 @@
 // Flutter imports:
 import 'dart:convert';
-import 'dart:ui' as ui;
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 
 // Package imports:
@@ -14,9 +12,9 @@ import 'package:shared_preferences/shared_preferences.dart';
 // Project imports:
 import 'package:binsight_ai/database/models/detection.dart';
 import 'package:binsight_ai/util/print.dart';
-import 'package:binsight_ai/util/providers.dart';
 import 'package:binsight_ai/widgets/free_draw.dart';
 import 'package:binsight_ai/widgets/heading.dart';
+import 'package:binsight_ai/util/providers/annotation_notifier.dart';
 
 /// Page used for annotating an individual detection image
 class AnnotationPage extends StatefulWidget {
