+++ lib/pages/setup/bluetooth.dart
@@ -10,8 +10,9 @@ import 'package:binsight_ai/util/async_ops.dart';
 import 'package:binsight_ai/util/bluetooth.dart';
 import 'package:binsight_ai/util/bluetooth_bin_data.dart';
 import 'package:binsight_ai/util/bluetooth_dialog_strings.dart';
+import 'package:binsight_ai/util/providers/device_notifier.dart';
+import 'package:binsight_ai/util/providers/setup_key_notifier.dart';
 import 'package:binsight_ai/util/print.dart';
-import 'package:binsight_ai/util/providers.dart';
 import 'package:binsight_ai/util/styles.dart';
 import 'package:binsight_ai/widgets/background.dart';
 import 'package:binsight_ai/widgets/bluetooth_alert_box.dart';
