+++ lib/pages/setup/index.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 // Project imports:
-import 'package:binsight_ai/util/providers.dart';
 import 'package:binsight_ai/widgets/background.dart';
+import 'package:binsight_ai/util/providers/setup_key_notifier.dart';
 
 /// The splash screen prompting the user to continue setting up their application.
 class SplashPage extends StatefulWidget {
