+++ lib/pages/setup/wifi_scan.dart
@@ -12,7 +12,9 @@ import 'package:binsight_ai/util/bluetooth.dart';
 import 'package:binsight_ai/util/bluetooth_dialog_strings.dart';
 import 'package:binsight_ai/util/bluetooth_exception.dart';
 import 'package:binsight_ai/util/print.dart';
-import 'package:binsight_ai/util/providers.dart';
+import 'package:binsight_ai/util/providers/device_notifier.dart';
+import 'package:binsight_ai/util/providers/setup_key_notifier.dart';
+import 'package:binsight_ai/util/providers/wifi_result_notifier.dart';
 import 'package:binsight_ai/util/styles.dart';
 import 'package:binsight_ai/util/wifi_scan.dart';
 import 'package:binsight_ai/widgets/background.dart';
