+++ lib/util/providers/setup_key_notifier.dart
@@ -1,0 +1,8 @@
+// Flutter imports:
+
+import 'package:flutter/material.dart';
+import 'package:introduction_screen/introduction_screen.dart';
+
+class SetupKeyNotifier {
+  final setupKey = GlobalKey<IntroductionScreenState>();
+}
