+++ lib/widgets/background.dart
@@ -24,9 +24,9 @@ class CustomBackground extends StatelessWidget {
   }
 }
 
-            // const ListTile(
-            //   title: Text(
-            //     'Version 1.0.0',
-            //     style: TextStyle(fontSize: 12, color: Colors.grey),
-            //   ),
-            // ),
+// const ListTile(
+//   title: Text(
+//     'Version 1.0.0',
+//     style: TextStyle(fontSize: 12, color: Colors.grey),
+//   ),
+// ),
